Refactored the technique descriptors.
Move the technique descriptors to their file and avoid exporting
internal symbols.
