+++ @here/harp-datasource-protocol/lib/Techniques.ts
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { Expr, JsonExpr } from "./Expr";
+import { InterpolatedPropertyDefinition } from "./InterpolatedPropertyDefs";
 import {
     BaseTechniqueParams,
     BasicExtrudedLineTechniqueParams,
@@ -13,7 +15,6 @@ import {
     LineTechniqueParams,
     MarkerTechniqueParams,
     PointTechniqueParams,
-    PolygonalTechniqueParams,
     SegmentsTechniqueParams,
     ShaderTechniqueParams,
     SolidLineTechniqueParams,
@@ -24,14 +25,6 @@ import {
     TextureCoordinateType
 } from "./TechniqueParams";
 
-import { Expr, JsonExpr } from "./Expr";
-import { InterpolatedPropertyDefinition } from "./InterpolatedPropertyDefs";
-import {
-    AttrScope,
-    mergeTechniqueDescriptor,
-    TechniqueDescriptor,
-    TechniqueDescriptorRegistry
-} from "./TechniqueDescriptor";
 /**
  * Names of the supported texture properties.
  * @internal
@@ -47,17 +40,11 @@ export const TEXTURE_PROPERTY_KEYS = [
     "bumpMap"
 ];
 
-/**
- * Names of the properties controlling transparency.
- * @internal
- */
-export const TRANSPARENCY_PROPERTY_KEYS = ["opacity", "transparent"];
-
 // TODO: Can be removed, when all when interpolators are implemented as {@link Expr}s
-export type RemoveInterpolatedPropDef<T> = T | InterpolatedPropertyDefinition<any> extends T
+type RemoveInterpolatedPropDef<T> = T | InterpolatedPropertyDefinition<any> extends T
     ? Exclude<T, InterpolatedPropertyDefinition<any>>
     : T;
-export type RemoveJsonExpr<T> = T | JsonExpr extends T ? Exclude<T, JsonExpr> : T;
+type RemoveJsonExpr<T> = T | JsonExpr extends T ? Exclude<T, JsonExpr> : T;
 
 /**
  * Make runtime representation of technique attributes from JSON-compatible typings.
@@ -71,435 +58,6 @@ export type MakeTechniqueAttrs<T> = {
         ? RemoveInterpolatedPropDef<RemoveJsonExpr<T[P]>> | Expr
         : T[P];
 };
-
-/** @internal  */
-export const techniqueDescriptors: TechniqueDescriptorRegistry = {};
-
-/** @internal  */
-export const baseTechniqueParamsDescriptor: TechniqueDescriptor<BaseTechniqueParams> = {
-    // TODO: Choose which techniques should support color with transparency.
-    // For now we chosen all, but it maybe not suitable for text or line marker techniques.
-    attrTransparencyColor: "color",
-    attrScopes: {
-        renderOrder: AttrScope.TechniqueGeometry,
-        enabled: AttrScope.FeatureGeometry,
-        kind: AttrScope.TechniqueGeometry,
-        transient: AttrScope.TechniqueGeometry,
-        fadeFar: AttrScope.TechniqueRendering,
-        fadeNear: AttrScope.TechniqueRendering
-    }
-};
-
-/** @internal  */
-export const pointTechniquePropTypes = mergeTechniqueDescriptor<PointTechniqueParams>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            texture: AttrScope.TechniqueGeometry,
-            enablePicking: AttrScope.TechniqueGeometry,
-            color: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueGeometry
-        }
-    }
-);
-
-/**
- * Runtime representation of [[SquaresStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[PointTechniqueParams]].
- */
-export interface SquaresTechnique extends MakeTechniqueAttrs<PointTechniqueParams> {
-    name: "squares";
-}
-
-/** @internal  */
-export const squaresTechniquePropTypes = mergeTechniqueDescriptor<SquaresTechnique>(
-    baseTechniqueParamsDescriptor,
-    pointTechniquePropTypes
-);
-techniqueDescriptors.squares = squaresTechniquePropTypes;
-
-/**
- * Runtime representation of [[CirclesStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[PointTechniqueParams]].
- */
-export interface CirclesTechnique extends MakeTechniqueAttrs<PointTechniqueParams> {
-    name: "circles";
-}
-
-/** @internal  */
-export const circlesTechniquePropTypes = mergeTechniqueDescriptor<CirclesTechnique>(
-    baseTechniqueParamsDescriptor,
-    pointTechniquePropTypes
-);
-techniqueDescriptors.circles = circlesTechniquePropTypes;
-
-/**
- * Runtime representation of [[PoiStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[MarkerTechniqueParams]].
- */
-export interface PoiTechnique extends MakeTechniqueAttrs<MarkerTechniqueParams> {
-    name: "labeled-icon";
-}
-
-/**
- * Runtime representation of [[LineMarkerStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[MarkerTechniqueParams]].
- */
-export interface LineMarkerTechnique extends MakeTechniqueAttrs<MarkerTechniqueParams> {
-    name: "line-marker";
-}
-
-const lineMarkerTechniquePropTypes = mergeTechniqueDescriptor<LineMarkerTechnique>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            text: AttrScope.FeatureGeometry,
-            label: AttrScope.FeatureGeometry,
-            useAbbreviation: AttrScope.FeatureGeometry,
-            useIsoCode: AttrScope.FeatureGeometry,
-            priority: AttrScope.TechniqueGeometry,
-            textMinZoomLevel: AttrScope.TechniqueGeometry,
-            textMaxZoomLevel: AttrScope.TechniqueGeometry,
-            iconMinZoomLevel: AttrScope.TechniqueGeometry,
-            iconMaxZoomLevel: AttrScope.TechniqueGeometry,
-            distanceScale: AttrScope.TechniqueGeometry,
-            textMayOverlap: AttrScope.TechniqueGeometry,
-            iconMayOverlap: AttrScope.TechniqueGeometry,
-            textReserveSpace: AttrScope.TechniqueGeometry,
-            iconReserveSpace: AttrScope.TechniqueGeometry,
-            renderTextDuringMovements: AttrScope.TechniqueGeometry,
-            alwaysOnTop: AttrScope.TechniqueGeometry,
-            textIsOptional: AttrScope.TechniqueGeometry,
-            showOnMap: AttrScope.TechniqueGeometry,
-            stackMode: AttrScope.TechniqueGeometry,
-            minDistance: AttrScope.TechniqueGeometry,
-            iconIsOptional: AttrScope.TechniqueGeometry,
-            iconFadeTime: AttrScope.TechniqueGeometry,
-            textFadeTime: AttrScope.TechniqueGeometry,
-            xOffset: AttrScope.TechniqueGeometry,
-            yOffset: AttrScope.TechniqueGeometry,
-            iconXOffset: AttrScope.TechniqueGeometry,
-            iconYOffset: AttrScope.TechniqueGeometry,
-            iconScale: AttrScope.TechniqueGeometry,
-            screenHeight: AttrScope.TechniqueGeometry,
-            screenWidth: AttrScope.TechniqueGeometry,
-            poiTable: AttrScope.TechniqueGeometry,
-            poiName: AttrScope.FeatureGeometry,
-            poiNameField: AttrScope.TechniqueGeometry,
-            imageTexture: AttrScope.FeatureGeometry,
-            imageTextureField: AttrScope.TechniqueGeometry,
-            imageTexturePrefix: AttrScope.TechniqueGeometry,
-            imageTexturePostfix: AttrScope.TechniqueGeometry,
-            iconColor: AttrScope.TechniqueRendering,
-            iconBrightness: AttrScope.TechniqueRendering,
-            style: AttrScope.TechniqueGeometry,
-            fontName: AttrScope.TechniqueGeometry,
-            fontStyle: AttrScope.TechniqueGeometry,
-            fontVariant: AttrScope.TechniqueGeometry,
-            rotation: AttrScope.TechniqueGeometry,
-            tracking: AttrScope.TechniqueGeometry,
-            leading: AttrScope.TechniqueGeometry,
-            maxLines: AttrScope.TechniqueGeometry,
-            lineWidth: AttrScope.TechniqueGeometry,
-            canvasRotation: AttrScope.TechniqueGeometry,
-            lineRotation: AttrScope.TechniqueGeometry,
-            wrappingMode: AttrScope.TechniqueGeometry,
-            hAlignment: AttrScope.TechniqueGeometry,
-            vAlignment: AttrScope.TechniqueGeometry,
-            placements: AttrScope.TechniqueGeometry,
-            backgroundColor: AttrScope.TechniqueRendering,
-            backgroundSize: AttrScope.TechniqueRendering,
-            backgroundOpacity: AttrScope.TechniqueRendering,
-            color: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            size: AttrScope.TechniqueRendering
-        }
-    }
-);
-techniqueDescriptors["line-marker"] = lineMarkerTechniquePropTypes;
-techniqueDescriptors["labeled-icon"] = lineMarkerTechniquePropTypes;
-
-/**
- * Runtime representation of [[SegmentsStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[SegmentsTechniqueParams]].
- */
-export interface SegmentsTechnique extends MakeTechniqueAttrs<SegmentsTechniqueParams> {
-    name: "segments";
-}
-
-const polygonalTechniqueDescriptor: TechniqueDescriptor<PolygonalTechniqueParams> = {
-    attrScopes: {
-        polygonOffset: AttrScope.TechniqueRendering,
-        polygonOffsetFactor: AttrScope.TechniqueRendering,
-        polygonOffsetUnits: AttrScope.TechniqueRendering,
-        lineColor: AttrScope.TechniqueRendering,
-        lineFadeFar: AttrScope.TechniqueRendering,
-        lineFadeNear: AttrScope.TechniqueRendering
-    }
-};
-/**
- * Runtime representation of [[BasicExtrudedLineStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[BasicExtrudedLineTechniqueParams]].
- */
-export interface BasicExtrudedLineTechnique
-    extends MakeTechniqueAttrs<BasicExtrudedLineTechniqueParams> {
-    name: "extruded-line";
-}
-
-/**
- * Runtime representation of [[StandardExtrudedLineStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[StandardExtrudedLineTechniqueParams]].
- */
-export interface StandardExtrudedLineTechnique
-    extends MakeTechniqueAttrs<StandardExtrudedLineTechniqueParams> {
-    name: "extruded-line";
-}
-
-/**
- * Runtime representation of [[SolidLineStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[SolidLineTechniqueParams]].
- */
-export interface SolidLineTechnique extends MakeTechniqueAttrs<SolidLineTechniqueParams> {
-    name: "solid-line" | "dashed-line";
-}
-
-/** @internal  */
-export const solidLineTechniqueDescriptor = mergeTechniqueDescriptor<SolidLineTechnique>(
-    baseTechniqueParamsDescriptor,
-    polygonalTechniqueDescriptor,
-    {
-        attrScopes: {
-            clipping: AttrScope.TechniqueGeometry,
-            secondaryRenderOrder: AttrScope.TechniqueGeometry,
-            color: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            lineWidth: AttrScope.TechniqueRendering,
-            secondaryWidth: AttrScope.TechniqueRendering,
-            secondaryColor: AttrScope.TechniqueRendering,
-            dashSize: AttrScope.TechniqueRendering,
-            gapSize: AttrScope.TechniqueRendering
-        }
-    }
-);
-techniqueDescriptors["solid-line"] = solidLineTechniqueDescriptor;
-// TODO: Remove deprecated "dashed-line" support in future releases.
-techniqueDescriptors["dashed-line"] = solidLineTechniqueDescriptor;
-
-/**
- * Runtime representation of [[LineStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[LineTechniqueParams]].
- */
-export interface LineTechnique extends MakeTechniqueAttrs<LineTechniqueParams> {
-    name: "line";
-}
-
-/** @internal  */
-export const lineTechniqueDescriptor = mergeTechniqueDescriptor<LineTechnique>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            // TODO, check, which are really dynamic !
-            color: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            lineWidth: AttrScope.FeatureGeometry
-        }
-    }
-);
-
-techniqueDescriptors.line = lineTechniqueDescriptor;
-
-/**
- * Runtime representation of [[FillStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[FillTechniqueParams]].
- */
-export interface FillTechnique extends MakeTechniqueAttrs<FillTechniqueParams> {
-    name: "fill";
-}
-
-const fillTechniqueDescriptor = mergeTechniqueDescriptor<FillTechnique>(
-    baseTechniqueParamsDescriptor,
-    polygonalTechniqueDescriptor,
-    {
-        attrScopes: {
-            color: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            lineWidth: AttrScope.TechniqueRendering
-        }
-    }
-);
-techniqueDescriptors.fill = fillTechniqueDescriptor;
-
-/**
- * Technique used to render a mesh geometry.
- * For technique parameters see [[StandardTechniqueParams]].
- */
-export interface StandardTechnique extends MakeTechniqueAttrs<StandardTechniqueParams> {
-    name: "standard";
-}
-const standardTechniqueDescriptor = mergeTechniqueDescriptor<StandardTechnique>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            color: AttrScope.FeatureGeometry,
-            vertexColors: AttrScope.FeatureGeometry,
-            wireframe: AttrScope.TechniqueRendering,
-            roughness: AttrScope.TechniqueRendering,
-            metalness: AttrScope.TechniqueRendering,
-            alphaTest: AttrScope.TechniqueRendering,
-            depthTest: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            emissive: AttrScope.TechniqueRendering,
-            emissiveIntensity: AttrScope.TechniqueRendering,
-            refractionRatio: AttrScope.TechniqueRendering,
-            map: AttrScope.TechniqueGeometry,
-            mapProperties: AttrScope.TechniqueGeometry,
-            normalMap: AttrScope.TechniqueGeometry,
-            normalMapProperties: AttrScope.TechniqueGeometry,
-            displacementMap: AttrScope.TechniqueGeometry,
-            displacementMapProperties: AttrScope.TechniqueGeometry,
-            roughnessMap: AttrScope.TechniqueGeometry,
-            roughnessMapProperties: AttrScope.TechniqueGeometry,
-            emissiveMap: AttrScope.TechniqueGeometry,
-            emissiveMapProperties: AttrScope.TechniqueGeometry,
-            bumpMap: AttrScope.TechniqueGeometry,
-            bumpMapProperties: AttrScope.TechniqueGeometry,
-            metalnessMap: AttrScope.TechniqueGeometry,
-            metalnessMapProperties: AttrScope.TechniqueGeometry,
-            alphaMap: AttrScope.TechniqueGeometry,
-            alphaMapProperties: AttrScope.TechniqueGeometry
-        }
-    }
-);
-techniqueDescriptors.standard = standardTechniqueDescriptor;
-
-/**
- * Runtime representation of [[ExtrudedPolygonStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[ExtrudedPolygonTechniqueParams]].
- */
-export interface ExtrudedPolygonTechnique
-    extends MakeTechniqueAttrs<ExtrudedPolygonTechniqueParams> {
-    name: "extruded-polygon";
-}
-
-const extrudedPolygonTechniqueDescriptor = mergeTechniqueDescriptor<ExtrudedPolygonTechnique>(
-    baseTechniqueParamsDescriptor,
-    standardTechniqueDescriptor,
-    {
-        attrScopes: {
-            height: AttrScope.FeatureGeometry,
-            floorHeight: AttrScope.FeatureGeometry,
-            color: AttrScope.FeatureGeometry,
-            defaultColor: AttrScope.FeatureGeometry,
-            defaultHeight: AttrScope.FeatureGeometry,
-            constantHeight: AttrScope.FeatureGeometry,
-            boundaryWalls: AttrScope.FeatureGeometry,
-            footprint: AttrScope.FeatureGeometry,
-            maxSlope: AttrScope.FeatureGeometry,
-            enableDepthPrePass: AttrScope.TechniqueGeometry,
-            animateExtrusionDuration: AttrScope.TechniqueGeometry,
-            animateExtrusion: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            transparent: AttrScope.TechniqueRendering,
-            lineWidth: AttrScope.TechniqueRendering,
-            lineFadeNear: AttrScope.TechniqueRendering,
-            lineFadeFar: AttrScope.TechniqueRendering,
-            lineColorMix: AttrScope.TechniqueGeometry,
-            lineColor: AttrScope.TechniqueRendering
-        }
-    }
-);
-techniqueDescriptors["extruded-polygon"] = extrudedPolygonTechniqueDescriptor;
-/**
- * Runtime representation of [[TextStyle]] as parsed by [[StyleSetEvaluator]].
- * For technique parameters see [[TextTechniqueParams]].
- */
-export interface TextTechnique extends MakeTechniqueAttrs<TextTechniqueParams> {
-    name: "text";
-}
-
-const textTechniqueDescriptor = mergeTechniqueDescriptor<TextTechnique>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            text: AttrScope.FeatureGeometry,
-            label: AttrScope.FeatureGeometry,
-            useAbbreviation: AttrScope.FeatureGeometry,
-            useIsoCode: AttrScope.FeatureGeometry,
-
-            minZoomLevel: AttrScope.TechniqueGeometry,
-            maxZoomLevel: AttrScope.TechniqueGeometry,
-            distanceScale: AttrScope.TechniqueGeometry,
-            mayOverlap: AttrScope.TechniqueGeometry,
-            reserveSpace: AttrScope.TechniqueGeometry,
-            textFadeTime: AttrScope.TechniqueGeometry,
-            xOffset: AttrScope.TechniqueGeometry,
-            yOffset: AttrScope.TechniqueGeometry,
-            style: AttrScope.TechniqueGeometry,
-            fontName: AttrScope.TechniqueGeometry,
-            fontStyle: AttrScope.TechniqueGeometry,
-            fontVariant: AttrScope.TechniqueGeometry,
-            rotation: AttrScope.TechniqueGeometry,
-            tracking: AttrScope.TechniqueGeometry,
-            leading: AttrScope.TechniqueGeometry,
-            maxLines: AttrScope.TechniqueGeometry,
-            lineWidth: AttrScope.TechniqueGeometry,
-            canvasRotation: AttrScope.TechniqueGeometry,
-            lineRotation: AttrScope.TechniqueGeometry,
-            wrappingMode: AttrScope.TechniqueGeometry,
-            hAlignment: AttrScope.TechniqueGeometry,
-            vAlignment: AttrScope.TechniqueGeometry,
-            backgroundColor: AttrScope.TechniqueRendering,
-            backgroundSize: AttrScope.TechniqueRendering,
-            backgroundOpacity: AttrScope.TechniqueRendering,
-            color: AttrScope.TechniqueRendering,
-            opacity: AttrScope.TechniqueRendering,
-            priority: AttrScope.TechniqueRendering,
-            size: AttrScope.TechniqueRendering
-        }
-    }
-);
-techniqueDescriptors.text = textTechniqueDescriptor;
-
-/**
- * Special technique for user-defined shaders.
- * For technique parameters see [[ShaderTechniqueParams]].
- */
-export interface ShaderTechnique extends MakeTechniqueAttrs<ShaderTechniqueParams> {
-    name: "shader";
-}
-
-const shaderTechniqueDescriptor = mergeTechniqueDescriptor<ShaderTechnique>(
-    baseTechniqueParamsDescriptor,
-    {
-        attrScopes: {
-            primitive: AttrScope.TechniqueGeometry,
-            params: AttrScope.TechniqueRendering
-        }
-    }
-);
-
-techniqueDescriptors.shader = shaderTechniqueDescriptor;
-
-/**
- * Technique used to render a terrain geometry with textures.
- * For technique parameters see [[TerrainTechniqueParams]].
- */
-export interface TerrainTechnique extends MakeTechniqueAttrs<TerrainTechniqueParams> {
-    name: "terrain";
-}
-
-/**
- * Technique to avoid label rendering on top of certain line geometries.
- * For technique parameters see [[BaseTechniqueParams]].
- */
-export interface LabelRejectionLineTechnique extends MakeTechniqueAttrs<BaseTechniqueParams> {
-    name: "label-rejection-line";
-}
 
 /**
  * Possible techniques that can be used to draw a geometry on the map.
@@ -523,36 +81,173 @@ export type Technique =
     | LabelRejectionLineTechnique;
 
 /**
+ * Runtime representation of `SquaresStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `PointTechniqueParams`.
+ */
+export interface SquaresTechnique extends MakeTechniqueAttrs<PointTechniqueParams> {
+    name: "squares";
+}
+
+/**
+ * Runtime representation of `CirclesStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `PointTechniqueParams`.
+ */
+export interface CirclesTechnique extends MakeTechniqueAttrs<PointTechniqueParams> {
+    name: "circles";
+}
+
+/**
+ * Runtime representation of `PoiStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `MarkerTechniqueParams`.
+ */
+export interface PoiTechnique extends MakeTechniqueAttrs<MarkerTechniqueParams> {
+    name: "labeled-icon";
+}
+
+/**
+ * Runtime representation of `LineMarkerStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `MarkerTechniqueParams`.
+ */
+export interface LineMarkerTechnique extends MakeTechniqueAttrs<MarkerTechniqueParams> {
+    name: "line-marker";
+}
+
+/**
+ * Runtime representation of `SegmentsStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `SegmentsTechniqueParams`.
+ */
+export interface SegmentsTechnique extends MakeTechniqueAttrs<SegmentsTechniqueParams> {
+    name: "segments";
+}
+
+/**
+ * Runtime representation of `BasicExtrudedLineStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `BasicExtrudedLineTechniqueParams`.
+ */
+export interface BasicExtrudedLineTechnique
+    extends MakeTechniqueAttrs<BasicExtrudedLineTechniqueParams> {
+    name: "extruded-line";
+}
+
+/**
+ * Runtime representation of `StandardExtrudedLineStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `StandardExtrudedLineTechniqueParams`.
+ */
+export interface StandardExtrudedLineTechnique
+    extends MakeTechniqueAttrs<StandardExtrudedLineTechniqueParams> {
+    name: "extruded-line";
+}
+
+/**
+ * Runtime representation of `SolidLineStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `SolidLineTechniqueParams`.
+ */
+export interface SolidLineTechnique extends MakeTechniqueAttrs<SolidLineTechniqueParams> {
+    name: "solid-line" | "dashed-line";
+}
+
+/**
+ * Runtime representation of `LineStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `LineTechniqueParams`.
+ */
+export interface LineTechnique extends MakeTechniqueAttrs<LineTechniqueParams> {
+    name: "line";
+}
+
+/**
+ * Runtime representation of `FillStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `FillTechniqueParams`.
+ */
+export interface FillTechnique extends MakeTechniqueAttrs<FillTechniqueParams> {
+    name: "fill";
+}
+
+/**
+ * Technique used to render a mesh geometry.
+ * For technique parameters see `StandardTechniqueParams`.
+ */
+export interface StandardTechnique extends MakeTechniqueAttrs<StandardTechniqueParams> {
+    name: "standard";
+}
+
+/**
+ * Runtime representation of `ExtrudedPolygonStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `ExtrudedPolygonTechniqueParams`.
+ */
+export interface ExtrudedPolygonTechnique
+    extends MakeTechniqueAttrs<ExtrudedPolygonTechniqueParams> {
+    name: "extruded-polygon";
+}
+
+/**
+ * Runtime representation of `TextStyle` as parsed by `StyleSetEvaluator`.
+ * For technique parameters see `TextTechniqueParams`.
+ */
+export interface TextTechnique extends MakeTechniqueAttrs<TextTechniqueParams> {
+    name: "text";
+}
+
+/**
+ * Special technique for user-defined shaders.
+ * For technique parameters see `ShaderTechniqueParams`.
+ */
+export interface ShaderTechnique extends MakeTechniqueAttrs<ShaderTechniqueParams> {
+    name: "shader";
+}
+
+/**
+ * Technique used to render a terrain geometry with textures.
+ * For technique parameters see `TerrainTechniqueParams`.
+ */
+export interface TerrainTechnique extends MakeTechniqueAttrs<TerrainTechniqueParams> {
+    name: "terrain";
+}
+
+/**
+ * Technique to avoid label rendering on top of certain line geometries.
+ * For technique parameters see `BaseTechniqueParams`.
+ */
+export interface LabelRejectionLineTechnique extends MakeTechniqueAttrs<BaseTechniqueParams> {
+    name: "label-rejection-line";
+}
+
+/**
+ * Names of the properties controlling transparency.
+ * @internal
+ */
+export const TRANSPARENCY_PROPERTY_KEYS = ["opacity", "transparent"];
+
+/**
  * Additional params used for optimized usage of `Techniques`.
  */
 export interface IndexedTechniqueParams {
     /**
-     * Optimization: Index into table in [[StyleSetEvaluator]] or in [[DecodedTile]].
+     * Optimization: Index into table in `StyleSetEvaluator` or in `DecodedTile`.
      * @hidden
      */
     _index: number;
 
     /**
-     * Optimization: Unique [[Technique]] index of [[Style]] from which technique was derived.
+     * Optimization: Unique `Technique` index of `Style` from which technique was derived.
      * @hidden
      */
     _styleSetIndex: number;
 
     /**
-     * The styleSet associated to this [[Technique]].
+     * The styleSet associated to this `Technique`.
      * @hidden
      */
     _styleSet?: string;
 
     /**
-     * The category used to assign render orders to objects created using this [[Technique]].
+     * The category used to assign render orders to objects created using this `Technique`.
      * @hidden
      */
     _category?: string;
 
     /**
      * The category used to assign render orders to secondary objects
-     * created using this [[Technique]].
+     * created using this `Technique`.
      * @hidden
      */
     _secondaryCategory?: string;
@@ -572,55 +267,55 @@ export interface IndexedTechniqueParams {
 }
 
 /**
- * For efficiency, [[StyleSetEvaluator]] returns [[Techniques]] additional params as defined in
- * [[IndexedTechniqueParams]].
+ * For efficiency, `StyleSetEvaluator` returns `Techniques` additional params as defined in
+ * `IndexedTechniqueParams`.
  */
 export type IndexedTechnique = Technique & IndexedTechniqueParams;
 
 /**
- * Type guard to check if an object is an instance of [[CirclesTechnique]].
+ * Type guard to check if an object is an instance of `CirclesTechnique`.
  */
 export function isCirclesTechnique(technique: Technique): technique is CirclesTechnique {
     return technique.name === "circles";
 }
 
 /**
- * Type guard to check if an object is an instance of [[SquaresTechnique]].
+ * Type guard to check if an object is an instance of `SquaresTechnique`.
  */
 export function isSquaresTechnique(technique: Technique): technique is SquaresTechnique {
     return technique.name === "squares";
 }
 
 /**
- * Type guard to check if an object is an instance of [[PoiTechnique]].
+ * Type guard to check if an object is an instance of `PoiTechnique`.
  */
 export function isPoiTechnique(technique: Technique): technique is PoiTechnique {
     return technique.name === "labeled-icon";
 }
 
 /**
- * Type guard to check if an object is an instance of [[LineMarkerTechnique]].
+ * Type guard to check if an object is an instance of `LineMarkerTechnique`.
  */
 export function isLineMarkerTechnique(technique: Technique): technique is LineMarkerTechnique {
     return technique.name === "line-marker";
 }
 
 /**
- * Type guard to check if an object is an instance of [[LineTechnique]].
+ * Type guard to check if an object is an instance of `LineTechnique`.
  */
 export function isLineTechnique(technique: Technique): technique is LineTechnique {
     return technique.name === "line";
 }
 
 /**
- * Type guard to check if an object is an instance of [[SolidLineTechnique]].
+ * Type guard to check if an object is an instance of `SolidLineTechnique`.
  */
 export function isSolidLineTechnique(technique: Technique): technique is SolidLineTechnique {
     return technique.name === "solid-line" || technique.name === "dashed-line";
 }
 
 /**
- * Type guard to check if an object is an instance of [[SolidLineTechnique]] and is a kind that
+ * Type guard to check if an object is an instance of `SolidLineTechnique` and is a kind that
  * has special dashes.
  * @note Lines with special dashes need line caps to render properly.
  */
@@ -635,15 +330,15 @@ export function isSpecialDashesLineTechnique(
 }
 
 /**
- * Type guard to check if an object is an instance of [[SegmentsTechnique]].
+ * Type guard to check if an object is an instance of `SegmentsTechnique`.
  */
 export function isSegmentsTechnique(technique: Technique): technique is SegmentsTechnique {
     return technique.name === "segments";
 }
 
 /**
- * Type guard to check if an object is an instance of [[BasicExtrudedLineTechnique]]
- * or [[StandardExtrudedLineTechnique]].
+ * Type guard to check if an object is an instance of `BasicExtrudedLineTechnique`
+ * or `StandardExtrudedLineTechnique`.
  */
 export function isExtrudedLineTechnique(
     technique: Technique
@@ -652,7 +347,7 @@ export function isExtrudedLineTechnique(
 }
 
 /**
- * Type guard to check if an object is an instance of [[BasicExtrudedLineTechnique]].
+ * Type guard to check if an object is an instance of `BasicExtrudedLineTechnique`.
  */
 export function isBasicExtrudedLineTechnique(
     technique: Technique
@@ -661,7 +356,7 @@ export function isBasicExtrudedLineTechnique(
 }
 
 /**
- * Type guard to check if an object is an instance of [[StandardExtrudedLineTechnique]].
+ * Type guard to check if an object is an instance of `StandardExtrudedLineTechnique`.
  */
 export function isStandardExtrudedLineTechnique(
     technique: Technique
@@ -670,14 +365,14 @@ export function isStandardExtrudedLineTechnique(
 }
 
 /**
- * Type guard to check if an object is an instance of [[FillTechnique]].
+ * Type guard to check if an object is an instance of `FillTechnique`.
  */
 export function isFillTechnique(technique: Technique): technique is FillTechnique {
     return technique.name === "fill";
 }
 
 /**
- * Type guard to check if an object is an instance of [[ExtrudedPolygonTechnique]].
+ * Type guard to check if an object is an instance of `ExtrudedPolygonTechnique`.
  */
 export function isExtrudedPolygonTechnique(
     technique: Technique
@@ -686,28 +381,28 @@ export function isExtrudedPolygonTechnique(
 }
 
 /**
- * Type guard to check if an object is an instance of [[StandardTechnique]].
+ * Type guard to check if an object is an instance of `StandardTechnique`.
  */
 export function isStandardTechnique(technique: Technique): technique is StandardTechnique {
     return technique.name === "standard";
 }
 
 /**
- * Type guard to check if an object is an instance of [[TerrainTechnique]].
+ * Type guard to check if an object is an instance of `TerrainTechnique`.
  */
 export function isTerrainTechnique(technique: Technique): technique is TerrainTechnique {
     return technique.name === "terrain";
 }
 
 /**
- * Type guard to check if an object is an instance of [[TextTechnique]].
+ * Type guard to check if an object is an instance of `TextTechnique`.
  */
 export function isTextTechnique(technique: Technique): technique is TextTechnique {
     return technique.name === "text";
 }
 
 /**
- * Type guard to check if an object is an instance of [[ShaderTechnique]].
+ * Type guard to check if an object is an instance of `ShaderTechnique`.
  */
 export function isShaderTechnique(technique: Technique): technique is ShaderTechnique {
     return technique.name === "shader";
